search/requirements.txt
numpy<2
streamlit>=1.10.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
watchdog
